cambios en la descarga del excel de muestras

+++ muestras/views.py
@@ -28,12 +28,15 @@ def muestras_todas(request):
     # Vista que muestra todas las muestras, requiere que el usuario esté autenticado
     muestras = Muestra.objects.prefetch_related('localizacion')
     # Filtrado de muestras si se proporcionan parámetros de búsqueda
-    field_names = [f.name for f in Muestra._meta.local_fields if f.name not in ('id')]
+    field_names = [f.name for f in Muestra._meta.local_fields if f.name not in ('id','estudio')]
     fields_loc = [f.name for f in Localizacion._meta.local_fields if f.name not in ('id','muestra')]
     for field in field_names:
         if request.GET.get(field):
             filter_kwargs = {f"{field}__icontains": request.GET[field]}
             muestras = muestras.filter(**filter_kwargs)
+    if request.GET.get('estudio'):
+        filtro_estudio = request.GET['estudio']
+        muestras = muestras.filter(estudio__nombre_estudio__icontains=filtro_estudio)
     '''
     # Crear un PDF con las muestras filtradas
     if request.GET.get('crear_pdf'):    
@@ -72,14 +75,22 @@ def muestras_todas(request):
             
             for field in field_names:
                 value = muestra.__dict__[field]
+                if value is None:
+                    value = ''
                 ws.cell(row_num, col_num).value= str(value)
                 col_num += 1
+            value = muestra.estudio.nombre_estudio if muestra.estudio else ''
+            ws.cell(row_num, col_num).value= str(value)
+            col_num += 1
             try:
                 loc = Localizacion.objects.get(muestra=muestra.nom_lab)
                 for field in fields_loc:
                     value = loc.__dict__[field]
+                    if value is None:
+                        value = ''
                     ws.cell(row_num, col_num).value= str(value)
                     col_num += 1
+                row_num += 1
             except:
                 row_num += 1
         wb.save(response)
@@ -150,7 +161,8 @@ def upload_excel(request):
         form = UploadExcel(request.POST, request.FILES)
         if 'confirmar' in request.POST:
             messages.success(request, 'Las muestras se han añadido correctamente.')
-            
+            ids_to_delete = request.session.pop('ids_error_muestras', [])
+            Muestra.objects.filter(nom_lab__in=ids_to_delete).delete()
         elif 'cancelar' in request.POST:
             # Eliminación de las muestras y localizaciones añadidas del excel
             ids_to_delete = request.session.pop('nuevos_ids', [])
@@ -250,15 +262,24 @@ def upload_excel(request):
                         estado_actual=normalize_value(row['estado_actual']),
                         estudio = estudio_instance
                         )
+                    def normalize_charfield_value(value):
+                        if pd.isna(value) or value is None:
+                            return None
+                        try:
+                            if isinstance(value, (float, int)) and value == int(value):
+                                return str(int(value))
+
+                        except ValueError:
+                            return str(value).strip()
                     localizacion, loc_created = Localizacion.objects.update_or_create(
                         muestra = muestra,
-                        congelador=row['congelador'],
-                        estante=row['estante'], 
-                        posicion_rack_estante=row['posicion_rack_estante'],
-                        rack=row['rack'],
-                        posicion_caja_rack=row['posicion_caja_rack'],
-                        caja=row['caja'],
-                        subposicion=row['subposicion']    
+                        congelador=normalize_charfield_value(row['congelador']),
+                        estante=normalize_charfield_value(row['estante']), 
+                        posicion_rack_estante=normalize_charfield_value(row['posicion_rack_estante']),
+                        rack=normalize_charfield_value(row['rack']),
+                        posicion_caja_rack=normalize_charfield_value(row['posicion_caja_rack']),
+                        caja=normalize_charfield_value(row['caja']),
+                        subposicion=normalize_value(row['subposicion'])    
                     )
                     if created:
                         nuevos_ids.append(muestra.id)
@@ -290,11 +311,13 @@ def upload_excel(request):
                                 elif column =='id_individuo':
                                     ids_error_muestras.append(row["nom_lab"])
                                     errors += 1
+                                    Muestra.objects.filter(id=nuevos_ids[len(nuevos_ids)-1]).delete()
                                 else:
                                     ids_error_muestras.append(row["nom_lab"])
                                     errors_loc+=1
-                                    localizacion.delete()
-                                Muestra.objects.filter(id=nuevos_ids[len(nuevos_ids)-1]).delete()
+                                    if not pd.isna(row['nom_lab']):
+                                        Muestra.objects.filter(id=nuevos_ids[len(nuevos_ids)-1]).delete()
+                                        localizacion.delete()
                             elif column in [f.name for f in Muestra._meta.local_fields if f.name not in ('nom_lab','id_individuo')]:
                                 Muestra.objects.filter(nom_lab=row['nom_lab']).update(**{column : None})
                                 ids_campos_vacios.append(row['nom_lab']) 
@@ -307,7 +330,7 @@ def upload_excel(request):
                         rack=row['rack'],
                         posicion_caja_rack=row['posicion_caja_rack'],
                         caja=row['caja'],
-                        subposicion=row['subposicion']
+                        subposicion=normalize_value(row['subposicion'])
                     ):
                         ids_localizaciones_ocupadas.append(row['nom_lab'])
                         localizaciones_ocupadas += 1
@@ -403,7 +426,7 @@ def descargar_plantilla(request,macro:int):
 def localizaciones(request):
     # Vista que muestra todas las localizaciones, tengan o no muestra
     localizaciones = Localizacion.objects.all().values().distinct()
-    congeladores = Localizacion.objects.exclude(congelador='').values_list('congelador', flat=True).distinct()
+    congeladores = Localizacion.objects.exclude(congelador=None).values_list('congelador', flat=True).distinct()
     
     estantes = (Localizacion.objects.exclude(estante='')
                 .values_list('congelador','estante')
@@ -697,7 +720,8 @@ def añadir_muestras_estudio(request):
             studio = Estudio.objects.get(nombre_estudio=study)
             muestras=Muestra.objects.filter(id__in=muestras)
             for muestra in muestras:
-                muestra.estudio.add(studio)
+                muestra.estudio = studio
+                muestra.save()
         if 'muestras_estudio' in request.session:
             del request.session['muestras_estudio']
         return redirect('muestras_todas')
@@ -706,6 +730,7 @@ def añadir_muestras_estudio(request):
 def repositorio_estudio(request, id_estudio):
     estudio = Estudio.objects.get(id_estudio=id_estudio)
     documentos = Documento.objects.filter(estudio = estudio, eliminado= False)
+    request.session['id_estudio'] = id_estudio
     # Filtrado opcional por usuario
     usuario = request.GET.get('usuario')
     if usuario:
@@ -740,11 +765,17 @@ def descargar_documento(request, documento_id,id_estudio):
     doc = Documento.objects.get(pk=documento_id, eliminado=False)      
     return FileResponse(open(doc.archivo.path, 'rb'), as_attachment=True, filename=os.path.basename(doc.archivo.name))
 
-def eliminar_documento(request, id_documento):
-    doc = Documento.objects.get(pk=id_documento, eliminado=False)
-    doc.eliminado = True
-    doc.fecha_eliminacion = timezone.now()
-    doc.save()
-    
+def eliminar_documento(request):
+    ids_documento = request.POST.getlist('doc_id')
+    for element in ids_documento:
+        try:
+            doc = Documento.objects.get(pk=element, eliminado=False)
+            doc.eliminado = True
+            doc.fecha_eliminacion = timezone.now()
+            doc.save()
+            return redirect('repositorio_estudio', id_estudio=doc.estudio)
+        except:
+            return redirect('repositorio_estudio', id_estudio=doc.estudio.id_estudio)
+    return redirect('repositorio_estudio', id_estudio=request.session.get('id_estudio'))
 # Vistas relacionadas con el envio de muestras
